Port the original CC repro consolidation
